pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-eval-harness"
version = "0.1.0"
description = "Agent and skill evaluation harness"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
mlflow = ["mlflow[genai]>=3.5"]
anthropic = ["anthropic>=0.40"]
anthropic-vertex = ["anthropic[vertex]>=0.40"]
all = [
    "mlflow[genai]>=3.5",
    "anthropic>=0.40",
]

[tool.setuptools.packages.find]
include = ["agent_eval*"]

